requirements.txt

python-telegram-bot==22.5
openai==2.4.0
python-dotenv==1.1.1
httpx==0.27.0